+++ app/editor/js/utils.js
@@ -2,9 +2,39 @@ export {
     Iterator,
     pathParse,
     saveAsFile,
+    merge, // 递归合并对象
 };
-import { getLocalFile } from './api.js';
 
+// REF [js - 对象递归合并merge - zc-lee - 博客园](https://www.cnblogs.com/zc-lee/p/15873611.html)
+function isObject(obj) {
+    return Object.prototype.toString.call(obj) === '[object Object]'
+}
+function isArray(arr) {
+    return Array.isArray(arr)
+}
+function merge(target, ...arg) {
+    return arg.reduce((acc, cur) => {
+        return Object.keys(cur).reduce((subAcc, key) => {
+            const srcVal = cur[key]
+            if (isObject(srcVal)) {
+                subAcc[key] = merge(subAcc[key] ? subAcc[key] : {}, srcVal)
+            } else if (isArray(srcVal)) {
+                // series: []，下层数组直接赋值
+                subAcc[key] = srcVal.map((item, idx) => {
+                    if (isObject(item)) {
+                        const curAccVal = subAcc[key] ? subAcc[key] : []
+                        return merge(curAccVal[idx] ? curAccVal[idx] : {}, item)
+                    } else {
+                        return item
+                    }
+                })
+            } else {
+                subAcc[key] = srcVal
+            }
+            return subAcc
+        }, acc)
+    }, target)
+}
 
 function* Iterator(items, loop = false) {
     // REF [ES6中的迭代器(Iterator)和生成器(Generator) - 小火柴的蓝色理想 - 博客园](https://www.cnblogs.com/xiaohuochai/p/7253466.html)